+++ lib/strava/models/authorization.rb
@@ -1,0 +1,7 @@
+module Strava
+  module Models
+    class Authorization < Model
+      property 'access_token'
+    end
+  end
+end